Moved StockData class into separate file
